Fix #8: Publish release artifacts to bintray
For now the process of publishing a new version is manual and described
in the README. This should be done automatically via a future
integration with `sbt-release`.

This also introduces a root project in order to exclude it from being
published and changes projects to use `moduleName` instead of `name`,
which seems more standard.