MPN-1558: fix failing test for object-RDD-DataFrame-RDD-object roundtrip
The type of the object we're trying to roundtrip shouldn't be an
instance member of a class; if it is, Spark will try to distribute the
parent object as well as the object we actually want to distribute. This
obviously fails because the parent object is a ScalaTest test object.